fix: move syft to sources package
