+++ internal/sources/syft/fixtures/alpine.cyclonedx
