+++ internal/sources/syft/fixtures/alpine.json
