+++ internal/sources/syft/fixtures/alpine.spdxjson
