+++ internal/sources/syft/fixtures/fedora.json
