+++ internal/sources/syft/fixtures/node.cyclonedx
