+++ internal/sources/syft/fixtures/node.json
