+++ internal/sources/syft/fixtures/node.spdxjson
