+++ internal/sources/syft/fixtures/redis.cyclonedx
