+++ internal/sources/syft/fixtures/redis.json
