+++ internal/sources/syft/fixtures/redis.spdxjson
