+++ internal/sources/syft/fixtures/redis_generated.json
