+++ internal/sources/syft/fixtures/redis_generated.spdxjson
