+++ internal/sources/syft/fixtures/update-fixtures.sh
