+++ internal/sources/syft/syft.go
